# libbeat: optimize asset data decoding