--- a/libbeat/asset/registry.go
+++ b/libbeat/asset/registry.go
@@ -21,8 +21,9 @@ import (
 	"bytes"
 	"compress/zlib"
 	"encoding/base64"
-	"io/ioutil"
 	"sort"
+
+	"github.com/elastic/elastic-agent-libs/iobuf"
 )
 
 // FieldsRegistry contains a list of fields.yml files
@@ -106,7 +107,6 @@ func EncodeData(data string) (string, error) {
 
 // DecodeData base64 decodes the data and uncompresses it
 func DecodeData(data string) ([]byte, error) {
-
 	decoded, err := base64.StdEncoding.DecodeString(data)
 	if err != nil {
 		return nil, err
@@ -119,5 +119,5 @@ func DecodeData(data string) ([]byte, error) {
 	}
 	defer r.Close()
 
-	return ioutil.ReadAll(r)
+	return iobuf.ReadAll(r)
 }