@Kavindu-Dodan 2025-01-02
It seems this introduced a build issue. I get `vet: fbreceiver/receiver_test.go:133:26: NewFactory().CreateLogsReceiver undefined (type receiver.Factory has no field or method CreateLogsReceiver)` [^1] from recent builds. 

[^1]: https://github.com/elastic/beats/actions/runs/12588586152/job/35086765916?pr=41817#step:8:216