@mauri870 2025-01-02
> It seems this introduced a build issue.

Yes, that is fixed in https://github.com/elastic/beats/pull/42199.